MB-57205: support connection_buffer_size > Integer.MAX_VALUE
Reviewed-on: https://review.couchbase.org/c/analytics-dcp-client/+/191946
Well-Formed: Restriction Checker
Tested-by: Michael Blow <michael.blow@couchbase.com>

+++ src/main/java/com/couchbase/client/dcp/util/FlowControlCallback.java
@@ -17,7 +17,7 @@ public interface FlowControlCallback {
     final FlowControlCallback NOOP = new FlowControlCallback() {
         @Override
         public void bufferAckWaterMarkReached(DcpAckHandle handle, DcpChannel dcpChannel, int ackCounter,
-                int ackWatermark) {
+                long ackWatermark) {
             // No Op
         }
 
@@ -27,7 +27,7 @@ public interface FlowControlCallback {
         }
     };
 
-    void bufferAckWaterMarkReached(DcpAckHandle handle, DcpChannel dcpChannel, int ackCounter, int ackWatermark);
+    void bufferAckWaterMarkReached(DcpAckHandle handle, DcpChannel dcpChannel, int ackCounter, long ackWatermark);
 
     void ackFlushedThroughNetwork(DcpAckHandle handle, DcpChannel dcpChannel);
